LRQA-39321 All jobs should be abled to do properties expansion when getting properties
